Replace 'baselines/compute_lesion_metrics.py' by 'correlations_with_clinical_scores/02_correlate_lesions_with_clinical_scores.py'
